+++ core/lib.rs
@@ -8,12 +8,11 @@ extern crate lazy_static;
 #[macro_use]
 extern crate log;
 
-mod any_error;
 mod bindings;
 mod core_isolate;
+mod errors;
 mod es_isolate;
 mod flags;
-mod js_errors;
 mod module_specifier;
 mod modules;
 mod ops;
@@ -24,11 +23,10 @@ mod zero_copy_buf;
 
 pub use rusty_v8 as v8;
 
-pub use crate::any_error::*;
 pub use crate::core_isolate::*;
+pub use crate::errors::*;
 pub use crate::es_isolate::*;
 pub use crate::flags::v8_set_flags;
-pub use crate::js_errors::*;
 pub use crate::module_specifier::*;
 pub use crate::modules::*;
 pub use crate::ops::*;
